Extend interface for Ganesha 2.0 recovery using dbus interface.
I believe that this interface covers all the needs from the 1.5 version. It is a
python script that takes event and IP or node-id as input. The only difference is
that if multiple IP or node-id needs to be taken over than the script should be
called multiple times, once for each IP. All events will put Ganesha in grace period.
In the case of EVEN_RELEASE_IP it will release all NLM locks for this IP, for now all
NLM locks are released. In the case of EVENT_TAKE_IP or EVENT_TAKE_NODEID it will read
in the NFSv4 client-id for the given IP or node-id.